% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aver_rule.R
\name{aver.rule}
\alias{aver.rule}
\title{Average Rule chart}
\usage{
aver.rule(data)
}
\arguments{
\item{data:}{Name of the dataset (must be in dataframe format).}
}
\description{
This function allows you to locate the number of dimensions which are important for CA interpretation, according to the so-called average rule. The reference line showing up in the returned histogram indicates the threshold of an optimal dimensionality of the solution according to the average rule.
}
\examples{
data(greenacre_data)
aver.rule(greenacre_data)

}
\keyword{average}
\keyword{rule}
