% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_coffee.r
\docType{data}
\name{brand_coffee}
\alias{brand_coffee}
\title{Dataset: Cross-tabulation of coffee brands vs. consumers' opinion}
\format{dataframe}
\usage{
data(brand_coffee)
}
\description{
Cross-tabulation (23x6) of the coffee brands against consumers' opinion.\cr
After: Kennedy R et al, Practical Applications of Correspondence Analysis to
Categorical Data in Market Research, in Journal of Targeting Measurement and
Analysis for Marketing, 1996
}
\keyword{datasets}
