% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakfast.r
\docType{data}
\name{breakfast}
\alias{breakfast}
\title{Dataset: Cross-tabulation of breakfast food vs consumers' opinion}
\format{
dataframe
}
\usage{
data(breakfast)
}
\description{
Cross-tabulation (14x8) of the breakfast food type against consumers'
opinion.\cr After: Bendixen M, A Practical Guide to the Use of Correspondence
Analysis in Marketing Research, in Research online 1, 1996, 16-38
}
\keyword{datasets}
