% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ca.aver.rule}
\alias{ca.aver.rule}
\title{Average Rule}
\usage{
\method{ca}{aver.rule}(data)
}
\arguments{
\item{data}{Name of the dataset.}
}
\description{
This function allows you to locate the number of dimensions which are important for CA interpretation, according to the so-called average rule. The reference line showing up in the returned histogram indicates the threshold of an optimal dimensionality of the solution according to the average rule.
}
\examples{
ca.aver.rule(data)
}
\keyword{average}
\keyword{rule}

