% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ca.cols.cntr}
\alias{ca.cols.cntr}
\title{Columns contribution}
\usage{
\method{ca}{cols.cntr}(data, num.dim)
}
\arguments{
\item{data}{Name of the dataset.}

\item{num.dim}{Number of the dimension for which the column categories contribution is returned.}
}
\description{
This function allows you to calculate the contribution of the column categories to the selected dimensions. The reference line showing up in the returned histogram indicates the threshold for a significant (i.e., important for the sake of dimension interpretation) contribution.
}
\examples{
ca.cols.cntr(data,1) Returns the contribution of the column categories to the 1 CA dimension.
}
\keyword{cols}
\keyword{contribution}

