% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ca.cols.corr}
\alias{ca.cols.corr}
\title{Columns correlation to the dimensions}
\usage{
\method{ca}{cols.corr}(data, num.dim)
}
\arguments{
\item{data}{Name of the dataset.}

\item{num.dim}{Number of the dimension for which the column categories correlation is returned.}
}
\description{
This function allows you to calculate the correlation of the column categories to the selected dimensions.
}
\examples{
ca.cols.corr(data, 1) Returns the correlation of the column categories to the 1 CA dimension.
}
\keyword{cols}
\keyword{correlation}

