% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ca.cols.qlt}
\alias{ca.cols.qlt}
\title{Columns quality of the display}
\usage{
\method{ca}{cols.qlt}(data, num.dim)
}
\arguments{
\item{data}{Name of the dataset.}

\item{num.dim}{Number dimension (minimum 2) for which the column categories quality is returned.}
}
\description{
This function allows you to calculate the quality of the display of the column categories on pairs of selected dimensions.
}
\examples{
ca.cols.qlt(data, 3) Returns the quality of the display of the column categories on the 1&2, 1&3 dimensions.
}
\keyword{cols}
\keyword{quality}

