% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ca.corr}
\alias{ca.corr}
\title{Correlation between rows and columns categories}
\usage{
\method{ca}{corr}(data)
}
\arguments{
\item{data}{Name of the dataset.}
}
\description{
This function allows you to calculate the strenght of the correlation between rows and columns of the contingency table. A reference line indicates the threshold above which the correlation can be considered important.
}
\examples{
ca.corr(data)
}
\keyword{contingency}
\keyword{correlation}
\keyword{table}

