% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ca.rows.cntr}
\alias{ca.rows.cntr}
\title{Rows contribution}
\usage{
\method{ca}{rows.cntr}(data, num.dim)
}
\arguments{
\item{data}{Name of the dataset.}

\item{num.dim}{Number of the dimension for which the row categories contribution is returned.}
}
\description{
This function allows you to calculate the contribution of the row categories to the selected dimensions. The reference line showing up in the returned histogram indicates the threshold for a significant (i.e., important for the sake of dimension interpretation) contribution.
}
\examples{
ca.rows.cntr(data, 1) Returns the contribution of the row categories to the 1 CA dimension.
}
\keyword{contribution}
\keyword{rows}

