% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ca.rows.corr}
\alias{ca.rows.corr}
\title{Rows correlation to the dimensions}
\usage{
\method{ca}{rows.corr}(data, num.dim)
}
\arguments{
\item{data}{Name of the dataset.}

\item{num.dim}{Number of the dimension for which the row categories correlation is returned.}
}
\description{
This function allows you to calculate the correlation of the row categories to the selected dimensions.
}
\examples{
ca.rows.corr(data, 1) Returns the correlation of the row categories to the 1 CA dimension.
}
\keyword{correlation}
\keyword{rows}

