% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ca.rows.qlt}
\alias{ca.rows.qlt}
\title{Rows quality of the display}
\usage{
\method{ca}{rows.qlt}(data, num.dim)
}
\arguments{
\item{data}{Name of the dataset.}

\item{num.dim}{Number dimension (minimum 2) for which the row categories quality is returned.}
}
\description{
This function allows you to calculate the quality of the display of the row categories on pairs of selected dimensions.
}
\examples{
ca.rows.qlt(data, 3) Returns the quality of the display of the row categories on the 1&2, 1&3 dimensions
}
\keyword{quality}
\keyword{rows}

