% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_scatter.r
\name{ca.scatter}
\alias{ca.scatter}
\title{Scatterplot visualization facility}
\usage{
\method{ca}{scatter}(data, x = 1, y = 2, type)
}
\arguments{
\item{data:}{name of the contingency table (must be in dataframe format).}

\item{x:}{first dimension to be plotted (x=1 by default).}

\item{y:}{second dimensions to be plotted (y=2 by default).}

\item{type:}{type of scatterplot to be returned (see examples).}
}
\description{
This function allows to get different types of CA scatterplots. It is just a wrapper for functions from the 'ca' and 'FactoMineR' packages.
}
\examples{
data(greenacre_data)
ca.scatter(greenacre_data, 1, 2, type=1) # symmetric scatterplot for rows and columns
ca.scatter(greenacre_data, 1, 2, type=2) # Standard Biplot; 2 plots are returned: one with row-categories vectors displayed, one for columns categories vectors. 
ca.scatter(greenacre_data, 1, 2, type=3) # scaterplot of row categories with groupings shown by different colors; scatterplot for column categories is also returned
ca.scatter(greenacre_data, 1, 2, type=4) # 3D scatterplot with cluster tree for row categories; scatterplot for column categories is also returned.

}
\keyword{analysis}
\keyword{correspondence}
\keyword{scatterplots}
