% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_corr.R
\name{caCorr}
\alias{caCorr}
\title{Chart of correlation between rows and columns categories}
\usage{
caCorr(data)
}
\arguments{
\item{data:}{name of the dataset (must be in dataframe format).}
}
\description{
This function allows you to calculate the strenght of the correlation between rows and columns of the contingency table. 
A reference line indicates the threshold above which the correlation can be considered important.
}
\examples{
data(greenacre_data)
ca.corr(greenacre_data)

}
\keyword{contingency}
\keyword{correlation}
\keyword{table}
