% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_corr.R
\name{cols.corr}
\alias{cols.corr}
\title{Chart of columns correlation with a selected dimension}
\usage{
cols.corr(data, x = 1, sort = TRUE, filter = FALSE, leg = TRUE,
  dotprightm = 5, cex.leg = 0.6, cex.labls = 0.75, leg.x.spc = 1,
  leg.y.spc = 1)
}
\arguments{
\item{data:}{name of the dataset (must be in dataframe format).}

\item{x:}{dimension for which the column categories correlation is returned (1st dimension by default).}

\item{sort:}{logical value (TRUE/FALSE) which allows to sort the categories in descending order of correlation with the selected dimension. TRUE is set by default.}

\item{filter:}{filter the row categories listed in the top-right legend, only showing those who have a major contribution to the definition of the selected dimension.}

\item{leg:}{enable (TRUE; default) or disable (FALSE) the legend at the right-hand side of the dotplot.}

\item{dotprightm:}{increases the empty space between the right margin of the dotplot and the left margin of the legend box.}

\item{cex.leg:}{adjust the size of the legend's characters.}

\item{cex.labls:}{adjust the size of the dotplot's labels.}

\item{leg.x.spc:}{adjust the horizontal space of the chart's legend. See more info from the 'legend' function's help (?legend).}

\item{leg.y.spc:}{adjust the y interspace of the chart's legend. See more info from the 'legend' function's help (?legend).}
}
\description{
This function allows to calculate the correlation (sqrt(COS2)) of the column categories with the selected dimension.
 
The function displays the correlation of the column categories with the selected dimension; the parameter sort=TRUE arrange the categories in decreasing order of correlation. 
At the left-hand side, the categories' labels show a symbol (+ or -) according to which side of the selected dimension they are correlated, either positive or negative. 
At the right-hand side, a legend (which is enabled/disabled using the 'leg' parameter) indicates the row categories' contribution (in permills) to the selected dimension (value enclosed within round brackets), and a symbol (+ or -) indicating whether they are actually contributing to the definition of the positive or negative side of the dimension, respectively. 
Further, an asterisk (*) flags the categories which can be considered major contributors to the definition of the dimension.
}
\examples{
data(greenacre_data)
cols.corr(greenacre_data, 1, sort=TRUE) #Plots the correlation of the column categories with the 1st CA dimension.

}
\keyword{cols}
\keyword{correlation}
