% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_loss.r
\docType{data}
\name{fire_loss}
\alias{fire_loss}
\title{Dataset: Cross-tabulation of cause of fire vs. amount of money loss}
\format{dataframe}
\usage{
data(fire_loss)
}
\description{
Cross-tabulation (9x4) of the amount of money loss against cause of fire.\cr
After: Li et al, Influences of Time, Location, and Cause Factors on the
Probability of Fire Loss in China: A Correspondence Analysis, in Fire
Technology 50(5), 2014, 1181-1200 (table 5)
}
\keyword{datasets}
