% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenacre_data.r
\docType{data}
\name{greenacre_data}
\alias{greenacre_data}
\title{Dataset: Cross-tabulation of funding category vs. University faculty}
\format{dataframe}
\usage{
data(greenacre_data)
}
\description{
Cross-tabulation (10x5) of funding category against University faculty.\cr
After: Greenacre M, Correspondence Analysis in Practice, Boca
Raton-London-New York, Chapman&Hall/CRC 2007 (exhibit 12.1)
}
\keyword{datasets}
