% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_cntr.R
\name{rows.cntr}
\alias{rows.cntr}
\title{Rows contribution chart}
\usage{
rows.cntr(data, x = 1, cti = FALSE, sort = TRUE, corr.thrs = 0,
  cex.labls = 0.75, dotprightm = 5, cex.leg = 0.6, leg.x.spc = 1,
  leg.y.spc = 1)
}
\arguments{
\item{data:}{name of the dataset (must be in dataframe format).}

\item{x:}{dimension for which the row categories contribution is returned (1st dimension by default).}

\item{cti:}{logical value (TRUE/FALSE) which specifies if the contribution to the total inertia must be displayed as well (FALSE by default).}

\item{sort:}{logical value (TRUE/FALSE) which allows to sort the categories in descending order of contribution to the inertia of the selected dimension. TRUE is set by default.}

\item{corr.thrs:}{threshold above which the column categories correlation will be displayed in the plot's legend.}

\item{cex.labls:}{adjust the size of the dotplot's labels.}

\item{dotprightm:}{increases the empty space between the right margin of the dotplot and the left margin of the legend box.}

\item{cex.leg:}{adjust the size of the legend's characters.}

\item{leg.x.spc:}{adjust the horizontal space of the chart's legend. See more info from the 'legend' function's help (?legend).}

\item{leg.y.spc:}{adjust the y interspace of the chart's legend. See more info from the 'legend' function's help (?legend).}
}
\description{
This function allows to calculate the contribution of the row categories to the selected dimension. 
It displays the contribution of the categories as a dotplot. A reference line indicates the threshold above which a contribution can be considered important for the determination of the selected dimension. 
The parameter cti=TRUE specifies that the categories' contribution to the total inertia is also shown (hollow circle). 
The parameter sort=TRUE sorts the categories in descending order of contribution to the inertia of the selected dimension. 
At the left-hand side of the plot, the categories' labels are given a symbol (+ or -) according to wheather each category is actually contributing to the definition of the positive or negative side of the dimension, respectively. 
At the right-hand side, a legend reports the correlation (sqrt(COS2)) of the column categories with the selected dimension. A symbol (+ or -) indicates with which side of the selected dimension each column category is correlated.
}
\examples{
data(greenacre_data)
rows.cntr(greenacre_data, 2, cti=TRUE, sort=TRUE) # Plots the contribution of the row categories to the 2nd CA dimension, and also displays the contribnution to the total inertia. The categories are sorted in descending order of contribution to the inertia of the selected dimension.

}
\keyword{contribution}
\keyword{rows}
