% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_qlt.R
\name{rows.qlt}
\alias{rows.qlt}
\title{Chart of rows quality of the display}
\usage{
rows.qlt(data, x = 1, y = 2, sort = TRUE, cex.labls = 0.75)
}
\arguments{
\item{data:}{name of the dataset (must be in dataframe format).}

\item{x:}{first dimension for which the quality is calculated (x=1 by default).}

\item{y:}{second dimension for which the quality is calculated (y=2 by default).}

\item{sort:}{logical value (TRUE/FALSE) which allows to sort the categories in descending order of quality of the representation on the subspace defined by the selected dimensions. TRUE is set by default.}

\item{cex.labls:}{adjust the size of the dotplot's labels.}
}
\description{
This function allows you to calculate the quality of the display of the row categories on pairs of selected dimensions.
}
\examples{
data(greenacre_data)
rows.qlt(greenacre_data,1,2,sort=TRUE) #Plots the quality of the display of the row categories on the 1&2 dimensions.

}
\keyword{quality}
\keyword{rows}
