% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{sig.ca.dim.perm}
\alias{sig.ca.dim.perm}
\title{Permuted significance of CA dimensions}
\usage{
sig.ca.dim.perm(data, x, y)
}
\arguments{
\item{data}{Name of the dataset.}

\item{x}{First dimension whose significance is calculated.}

\item{y}{Second dimension whose significance is calculated.}
}
\description{
This function allows you to calculate the permuted significance of selected CA dimensions. Number of permutation set at 999. Since the function returns a scatterplot, the function requires that two dimensions are entered.
}
\examples{
sig.ca.dim.perm(data, 1,2) Returns a scatterplot of the permuted inertia of the 1 CA dimension against the permuted inertia of the 2 CA dimension. Observed inertia of the selected dimensions and 95th percentile of the permuted inertias are also displayed for testing the significance of the observed inertias.
}
\keyword{dimensions}
\keyword{permuted}
\keyword{significance}

