% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_dim_perm.R
\name{sig.dim.perm}
\alias{sig.dim.perm}
\title{Permuted significance of CA dimensions}
\usage{
sig.dim.perm(data, x = 1, y = 2, B = 1000)
}
\arguments{
\item{data:}{name of the dataset (must be in dataframe format).}

\item{x:}{first dimension whose significance is calculated (x=1 by default).}

\item{y:}{second dimension whose significance is calculated (y=2 by default).}

\item{B:}{number of permutations (999 by default).}
}
\description{
This function allows you to calculate the permuted significance of selected CA dimensions. Number of permutation set at 999 by default, but can be increased by the user. 
Since the function returns a scatterplot, the function requires that two dimensions are entered.
}
\examples{
data(greenacre_data)
sig.dim.perm(greenacre_data, 1,2) #Returns a scatterplot of the permuted inertia of the 1 CA dimension against the permuted inertia of the 2 CA dimension. Observed inertia of the selected dimensions and 95th percentile of the permuted inertias are also displayed for testing the significance of the observed inertias.

}
\keyword{dimensions}
\keyword{permuted}
\keyword{significance}
