% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_corr.R
\name{caCorr}
\alias{caCorr}
\title{Chart of correlation between rows and columns categories}
\usage{
caCorr(data)
}
\arguments{
\item{data}{Name of the dataset (in dataframe format).}
}
\description{
This function calculates the strength of the correlation between
rows and columns of the contingency table. A reference line indicates the
threshold above which the correlation can be considered important.
}
\examples{
data(greenacre_data)
caCorr(greenacre_data)

}
\keyword{caCorr}
