\name{CAinterprTools-package}
\alias{CAinterprTools-package}
\alias{CAinterprTools}
\docType{package}
\title{
Package for Graphical Aid in Correspondence Analysis Interpretation and Significance Testings
}
\description{
CAinterTools is a package that allows to plot a number of information related to the 
interpretation of Correspondence Analysis' results. It provides the facility to plot the 
contribution of rows and columns categories to the principal dimensions, the quality of points 
display on selected dimensions, the correlation of row and column categories to 
selected dimensions, etc. It also allows to assess which dimension(s) is important for the data 
structure interpretation by means of different statistics and tests. The package also offers the 
facility to plot the permuted distribution of the table total inertia as well as of the inertia 
accounted for by pairs of selected dimensions. Different facilities are also provided that aim 
to produce interpretation-oriented scatteplots.
}
\details{
\tabular{ll}{
Package: \tab CAinterprTools\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2018-05\cr
License: \tab GPL\cr
}
}
\author{
Gianmarco ALBERTI

Maintainer: Gianmarco ALBERTI <gianmarco.alberti@um.edu.mt>
}
\references{
Alberti G. 2013, An R script to facilitate Correspondence Analysis. A guide to the use and the interpretation of results from an archaeological perspective, Archeologia e Calcolatori 24, 25-54.

Alberti G. 2015, CAinterprTools: An R package to help interpreting Correspondence Analysis' results, SoftwareX 1-2, 26-31.

Benzecri J.P. 1992, Correspondence Analysis Handbook, New York, Marcel Dekker.

Blasius J., Greenacre M. 1998, Visualization of Categorical Data, San Diego-London, Academic Press.

Greenacre M. 2007, Correspondence Analysis in Practice, Boca Raton-London-New York, Chapman&Hall/CRC.

Le S., Josse J., Husson F. 2008, FactoMineR: An R package for multivariate analysis, Journal of Statistical Software, 25, 1-18.

Nenadic O., Greenacre M. 2007, Correspondence Analysis in R, with two- and three-dimensional graphics: The ca package, Journal of Statistical Software, 20, 1-13.

Saporta G. 2006, Probabilites, analyse des donnees et statistique (2e ed.), Paris, Editions Technip.
}
\keyword{ package }
\seealso{
ca,FactoMineR
}
