% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{malinvaud}
\alias{malinvaud}
\title{Malinvaud's test for significance of the CA dimensions}
\usage{
malinvaud(data)
}
\arguments{
\item{data}{Name of the dataset.}
}
\description{
This function allows you to perform the Malinvaud's test, which assesses the significance of the CA dimensions. The function returns both a table in the R console and a chart. The former lists some values, among which the significance of each CA dimension. The chart is a scatterplot of dimensions against p value. Dimensions whose p value is below the 0.05 threshold (displayed in RED) are significant.
}
\examples{
malinvaud(data)
}
\keyword{Malinvaud}
\keyword{test}

